src/gtfs/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.1.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
decorator==4.4.2
fsspec==0.8.0
gcsfs==0.7.0
google==3.0.0
google-api-core==1.22.1
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gtfs-realtime-bindings==0.0.6
idna==2.10
multidict==4.7.6
numpy==1.19.1
oauthlib==3.1.0
pandas==1.1.1
protobuf==3.13.0
protobuf3-to-dict==0.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
yarl==1.5.1
